Turned error logging middleware into responce logging one
