+++ VictoryCenter/VictoryCenter.WebAPI/Extensions/ApplicatonConfiguration.cs
@@ -4,9 +4,9 @@ namespace VictoryCenter.WebAPI.Extensions;
 
 public static class ApplicationConfiguration
 {
-    public static void UseErrorLogging(this IApplicationBuilder app)
+    public static void UseRequestResponseLogging(this IApplicationBuilder app)
     {
         app.UseMiddleware<ExceptionHandlingMiddleware>();
-        app.UseMiddleware<ErrorLoggingMiddleware>();
+        app.UseMiddleware<RequestResponseLoggingMiddleware>();
     }
 }
